Move all code to one class
