+++ Inviter/Inviter.cs
@@ -21,7 +21,6 @@ using Dalamud.Hooking;
 using Dalamud.Game.Internal.Network;
 using GroupManager = Inviter.ClientStructs.GroupManager;
 using PartyMember = Inviter.ClientStructs.PartyMember;
-using Dalamud.Game.Internal.Gui.Toast;
 
 namespace Inviter
 {
@@ -49,7 +48,7 @@ namespace Inviter
         private Int64 uiInvite;
         private IntPtr groupManagerAddress;
         private Dictionary<string, Int64> name2CID;
-        internal ThreadTimedEnable timedRecruitment;
+        internal TimedEnable timedRecruitment;
 
         public void Dispose()
         {
@@ -112,7 +111,7 @@ namespace Inviter
             Interface.Framework.Gui.Chat.OnChatMessage += Chat_OnChatMessage;
             // Interface.Framework.Network.OnNetworkMessage += Chat_OnNetworkMessage;
             Interface.ClientState.TerritoryChanged += TerritoryChanged;
-            timedRecruitment = new ThreadTimedEnable(this);
+            timedRecruitment = new TimedEnable(this);
         }
 
         private void TerritoryChanged(object sender, ushort e)
@@ -154,42 +153,7 @@ namespace Inviter
             }
             else
             {
-                if (Config.Enable && !timedRecruitment.isRunning)
-                {
-                    Interface.Framework.Gui.Toast.ShowError("Can't start timed recruitment because Inviter is turned on permanently");
-                }
-                else
-                {
-                    try
-                    {
-                        var time = int.Parse(args);
-                        if (time > 0)
-                        {
-                            timedRecruitment.runUntil = DateTimeOffset.Now.ToUnixTimeSeconds() + time * 60;
-                            if (!timedRecruitment.isRunning)
-                            {
-                                new Thread(new ThreadStart(timedRecruitment.Run)).Start();
-                            }
-                            Interface.Framework.Gui.Toast.ShowQuest("Commenced automatic recruitment for " + time + " minutes", new QuestToastOptions()
-                            {
-                                DisplayCheckmark = true,
-                                PlaySound = true
-                            });
-                        }
-                        else if (time == 0)
-                        {
-                            timedRecruitment.runUntil = 0;
-                        }
-                        else
-                        {
-                            Interface.Framework.Gui.Toast.ShowError("Time can not be negative");
-                        }
-                    }
-                    catch (Exception)
-                    {
-                        Interface.Framework.Gui.Toast.ShowError("Please enter amount of time in minutes");
-                    }
-                }
+                timedRecruitment.ProcessCommandTimedEnable(args);
             }
             /*
             else if (args == "party")
